error handling

+++ Frontend/src/Components/Registration/ChangeScheduling.jsx
@@ -60,11 +60,12 @@ export default function ChangeScheduling() {
     setOpenConfirmationDialog(true);
     setSubmitting(false);
   };
-  console.log(scheID);
   useEffect(() => {
     const id = sessionStorage.getItem("id");
     setScheID(id);
   }, [Error]);
+
+  console.log(scheID, "<======== ID")
 
   function generateTimeOptions(intervalMinutes) {
     const timeOptions = [];
@@ -79,11 +80,10 @@ export default function ChangeScheduling() {
   }
 
   const timeOptions = generateTimeOptions(15);
-  console.log(timeOptions);
 
   const handleConfirmation = async () => {
     set_loading(true);
-    console.log(Form_values);
+
     try {
       const response = await axios.post(
         `${process.env.REACT_APP_BASE_URL}/schedule/create-new-schedule`,
@@ -101,26 +101,27 @@ export default function ChangeScheduling() {
           },
         }
       );
+
+      if (response.data.message === false) {
+        if (scheID === null || scheID === "undefined") {
+          sessionStorage.setItem("id", response?.data?.data?.data?.id);
+        }
+        set_Error(response.data?.data?.data);
+        set_openError(true);
+      } else {
+        setOpenDialog(true);
+        set_success(response.data?.data);
+        sessionStorage.clear("id");
+        setScheID(null);
+      }
       set_loading(false);
-      console.log(response);
-      set_success(response.data?.data);
-      sessionStorage.clear("id");
-      setScheID(null);
-      setOpenDialog(true);
     } catch (error) {
       set_loading(false);
-      if (scheID === null) {
-        sessionStorage.setItem("id", error.response?.data?.data?.id);
-      }
-
-      set_Error(error.response.data);
-      set_openError(true);
     }
     setOpenConfirmationDialog(false);
   };
 
   const Update_Schedule = async (id) => {
-    console.log("Updated Schedule button clicked");
     try {
       const response = await axios.put(
         `${process.env.REACT_APP_BASE_URL}/schedule/update-schedule`,
@@ -469,7 +470,9 @@ export default function ChangeScheduling() {
               <h2>Something went wrong</h2>
             </div>
             <div className="flex mt-5 flex-row justify-around">
-              <div className="flex ml-2 text-[#4A4545]">{Error?.message}</div>
+              <div className="flex ml-2 text-[#4A4545]">
+                {Error?.validation_error}
+              </div>
             </div>
           </div>
         </DialogContent>
